pyproject.toml
[project]
name = "1-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.5.0",
    "mcp>=1.7.1",
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.2",
]
